Move fetch location into relevant component; ideally this makes ObsEdit stop rerendering every second
