+++ src/components/ObsEdit/EvidenceSection.js
@@ -4,9 +4,12 @@ import PhotoCarousel from "components/SharedComponents/PhotoCarousel";
 import { Text, View } from "components/styledComponents";
 import { UploadContext } from "providers/contexts";
 import type { Node } from "react";
-import React, { useContext } from "react";
+import React, {
+  useContext, useEffect, useRef, useState
+} from "react";
 import { useTranslation } from "react-i18next";
 import { createObservedOnStringForUpload } from "sharedHelpers/dateAndTime";
+import fetchUserLocation from "sharedHelpers/fetchUserLocation";
 
 import DatePicker from "./DatePicker";
 
@@ -16,6 +19,9 @@ type Props = {
   handleAddEvidence?: Function
 }
 
+const INITIAL_POSITIONAL_ACCURACY = 99999;
+const TARGET_POSITIONAL_ACCURACY = 10;
+
 const EvidenceSection = ( {
   handleSelection,
   handleAddEvidence,
@@ -23,8 +29,14 @@ const EvidenceSection = ( {
 }: Props ): Node => {
   const {
     currentObs,
-    updateObservationKey
+    updateObservationKey,
+    updateObservationKeys
   } = useContext( UploadContext );
+  const mountedRef = useRef( true );
+  const [shouldFetchLocation, setShouldFetchLocation] = useState( !currentObs?._created_at );
+  const [fetchingLocation, setFetchingLocation] = useState( false );
+  const [positionalAccuracy, setPositionalAccuracy] = useState( INITIAL_POSITIONAL_ACCURACY );
+
   const { t } = useTranslation( );
 
   const formatDecimal = coordinate => coordinate && coordinate.toFixed( 6 );
@@ -33,6 +45,76 @@ const EvidenceSection = ( {
 
   const latitude = currentObs?.latitude;
   const longitude = currentObs?.longitude;
+
+  // Hook version of componentWillUnmount. We use a ref to track mounted
+  // state (not useState, which might get frozen in a closure for other
+  // useEffects), and set it to false in the cleanup cleanup function. The
+  // effect has an empty dependency array so it should only run when the
+  // component mounts and when it unmounts, unlike in the cleanup effects of
+  // other hooks, which will run when any of there dependency values change,
+  // and maybe even before other hooks execute. If we ever need to do this
+  // again we could probably wrap this into its own hook, like useMounted
+  // ( ).
+  useEffect( ( ) => {
+    mountedRef.current = true;
+    return function cleanup( ) {
+      mountedRef.current = false;
+    };
+  }, [] );
+
+  useEffect( ( ) => {
+    if ( !currentObs ) return;
+
+    if ( !shouldFetchLocation ) return;
+
+    if ( fetchingLocation ) return;
+
+    const fetchLocation = async () => {
+      // If the component is gone, you won't be able to updated it
+      if ( !mountedRef.current ) return;
+
+      if ( !shouldFetchLocation ) return;
+      setFetchingLocation( false );
+
+      const location = await fetchUserLocation( );
+
+      // If we're still receiving location updates and location is blank,
+      // then we don't know where we are any more and the obs should update
+      // to reflect that
+      updateObservationKeys( {
+        place_guess: location?.place_guess,
+        latitude: location?.latitude,
+        longitude: location?.longitude,
+        positional_accuracy: location?.positional_accuracy
+      } );
+
+      // The local state version of positionalAccuracy needs to be a number,
+      // so don't set it to
+      const newPositionalAccuracy = location?.positional_accuracy || INITIAL_POSITIONAL_ACCURACY;
+      setPositionalAccuracy( newPositionalAccuracy );
+    };
+
+    if (
+      // If we're already fetching we don't need to fetch again
+      !fetchingLocation
+      // We only need to fetch when we're above the target
+      && positionalAccuracy >= TARGET_POSITIONAL_ACCURACY
+    ) {
+      setFetchingLocation( true );
+      // No need to fetch more than once a second
+      setTimeout( fetchLocation, 1000 );
+    } else {
+      setShouldFetchLocation( false );
+    }
+  }, [
+    currentObs,
+    fetchingLocation,
+    positionalAccuracy,
+    setFetchingLocation,
+    setShouldFetchLocation,
+    shouldFetchLocation,
+    updateObservationKeys
+  ] );
 
   const handleDatePicked = selectedDate => {
     if ( selectedDate ) {
